requirements.txt
django=4.2.3
djangorestframework=3.14.0
django-filter=23.2
djoser=2.2.0
gunicorn=20.1.0
psycopg2-binary=2.9.6
python-dotenv=1.0.0
Pillow=10.0.0
